refactor: put extended and custom component into 1 folder
