+++ app/globals.css
@@ -8,6 +8,21 @@ body,
   height: 100%;
 }
 
+/* all browser scrollbar, except firefox */
+::-webkit-scrollbar {
+  width: 4px;
+}
+
+::-webkit-scrollbar-thumb {
+  background-color: #8c8c8c;
+  border-radius: 8px;
+}
+
+::-webkit-scrollbar-track {
+  background-color: #f1f1f1;
+  border-radius: 4px;
+}
+
 @layer base {
   :root {
     --background: 0 0% 100%;
