+++ components/chat/chat-header.tsx
@@ -1,8 +1,8 @@
 import { Hash } from 'lucide-react'
 
+import MobileToggle from '@/components/feature/mobile-toggle'
 import SocketIndicator from '@/components/feature/socket-indicator'
-import MobileToggle from '@/components/mobile-toggle'
-import UserAvatar from '@/components/user-avatar'
+import UserAvatar from '@/components/feature/user-avatar'
 import { getInitials } from '@/lib/function'
 import { EChatHeaderType } from '@/types/enums'
