+++ components/modals/edit-server-modal.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/navigation'
 import { useEffect } from 'react'
 import { useForm } from 'react-hook-form'
 
-import FileUpload from '@/components/file-upload'
+import FileUpload from '@/components/feature/file-upload'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
