+++ components/modals/initial-modal.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation'
 import { useEffect, useState } from 'react'
 import { useForm } from 'react-hook-form'
 
-import FileUpload from '@/components/file-upload'
+import FileUpload from '@/components/feature/file-upload'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
